28.05.2014 12:26

+++ labor5/CEPtemplate/Project/spi_mem.h
@@ -5,6 +5,6 @@
 
 void spi_mem_EraseChip(unsigned int chip_sel);
 unsigned int spi_mem_ReadManufacturerID(unsigned int chip_sel);
-uint16_t spi_mem_ReadManufacturerID_Heitmann(void);
+uint32_t spi_mem_ReadManufacturerID_Heitmann(void);
 
 #endif /* SPI_MEM_ */
